Fixed ability to restrict the data shown in the
ScatterplotSimulationPlotter and TimeSeriesSimulationPlotter objects to
the most recent 'maxSamples' updates.  Also created the ability to
disable the chart legend in those two plotters.  Introduced code for
automatic updating of GUI parameters, but commented out as issues to do
with handling bad parameter values (i.e. null values) are not resolved.